Move the postgis_jts Driver class and some other into a new bundle which is and H2 fragment.
