+++ bundles/postgis-jts-osgi/pom.xml
@@ -51,6 +51,16 @@
             <artifactId>postgis-jdbc-jtsparser</artifactId>
             <version>2.1.0SVN</version>
         </dependency>
+        <dependency>
+            <groupId>org.orbisgis</groupId>
+            <artifactId>postgis-jts</artifactId>
+            <version>5.1.0-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.orbisgis</groupId>
+            <artifactId>postgis-jts</artifactId>
+            <version>5.1.0-SNAPSHOT</version>
+        </dependency>
     </dependencies>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
