+++ bundles/postgis-jts/src/main/java/org/orbisgis/postgis_jts/Driver.java
@@ -1,4 +1,4 @@
-package org.orbisgis.postgis_jts_osgi;
+package org.orbisgis.postgis_jts;
 
 import org.postgis.jts.JtsWrapper;
 
@@ -10,7 +10,7 @@ import java.util.logging.Level;
 /**
  * Driver that provide directly the JTS Object when calling {@link java.sql.ResultSet#getObject(int)}
  * To create linked table on H2 database to PostGIS db:
- * CREATE LINKED TABLE mytable('org.orbisgis.postgis_jts_osgi.Driver',
+ * CREATE LINKED TABLE mytable('org.orbisgis.postgis_jts.Driver',
  * 'jdbc:postgresql_h2://serverdomain:5432/databasename', 'user', 'password', '(select * from mytable)');
  * @author Nicolas Fortin
  */
