+++ bundles/postgis-jts/src/main/java/org/orbisgis/postgis_jts/ResultSetMetaDataWrapper.java
@@ -26,7 +26,7 @@
  * or contact directly:
  * info_at_ orbisgis.org
  */
-package org.orbisgis.postgis_jts_osgi;
+package org.orbisgis.postgis_jts;
 
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
